Added fix for a crawlspider bug
When there is a crawlspider with multiple rules, if one of them has no methods attached as a callback, then it gets "TypeError: join() argument must be str, bytes, or os.PathLike object, not 'NoneType'". It tries to make a directory based on the name, but fails to do so. Adding a check prevents this issue